apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.3
scipy==1.11.4
reportlab==4.0.9
websockets==12.0
email-validator==2.1.0
bcrypt==4.0.1
pywebview==5.0.1
google-generativeai==0.8.3
psycopg2-binary==2.9.9

